Protect against unset $HOME in ~/.profile

+++ shell/shared/.profile
@@ -10,7 +10,7 @@ fi
 if command -v ssh-agent >/dev/null 2>&1; then
 
     ssh_connect() {
-        if [ -f "$HOME/.ssh/agent-info" ]; then
+        if [ -n "$HOME" ] && [ -f "$HOME/.ssh/agent-info" ]; then
             eval "$(cat "$HOME/.ssh/agent-info")" >/dev/null
         fi
     }
@@ -24,10 +24,12 @@ if command -v ssh-agent >/dev/null 2>&1; then
     }
 
     ssh_restart() {
-        pkill ssh-agent
-        mkdir -p "$HOME/.ssh"
-        ssh-agent -t 86400 > "$HOME/.ssh/agent-info"
-        ssh_connect
+        if [ -n "$HOME" ]; then
+            pkill ssh-agent
+            mkdir -p "$HOME/.ssh"
+            ssh-agent -t 86400 > "$HOME/.ssh/agent-info"
+            ssh_connect
+        fi
     }
 
     ssh_connect
